+++ src/core/application.rs
@@ -307,6 +307,7 @@ where
                 .map(|x| x.map(|x| vec![x]).unwrap_or_default()),
             PollStrategy::TryFor(timeout) => self.poll_try_for(timeout),
             PollStrategy::UpTo(times) => self.poll_upto(times),
+            PollStrategy::UpToNoWait(times) => self.poll_upto_nowait(times),
         }
     }
 
@@ -315,6 +316,32 @@ where
         let mut evs: Vec<Event<UserEvent>> = Vec::with_capacity(t);
         for _ in 0..t {
             match self.poll_listener_timeout() {
+                Err(err) => return Err(err),
+                Ok(None) => break,
+                Ok(Some(ev)) => evs.push(ev),
+            }
+        }
+        Ok(evs)
+    }
+
+    /// Poll event listener up to `t` times, without waiting to return if there are events.
+    fn poll_upto_nowait(&mut self, t: usize) -> ApplicationResult<Vec<Event<UserEvent>>> {
+        if t == 0 {
+            return Ok(Vec::new());
+        }
+
+        let mut evs: Vec<Event<UserEvent>> = Vec::with_capacity(t);
+
+        match self.poll_listener_timeout() {
+            Err(err) => return Err(err),
+            Ok(None) => (),
+            Ok(Some(ev)) => evs.push(ev),
+        }
+
+        let t = t.saturating_sub(1);
+
+        for _ in 0..t {
+            match self.try_poll_listener() {
                 Err(err) => return Err(err),
                 Ok(None) => break,
                 Ok(Some(ev)) => evs.push(ev),
@@ -340,6 +367,11 @@ where
     /// Poll event listener once with timeout
     fn poll_listener_timeout(&mut self) -> ApplicationResult<Option<Event<UserEvent>>> {
         self.listener.poll_timeout().map_err(ApplicationError::from)
+    }
+
+    /// Try to Poll event listener once, without blocking whatsoever
+    fn try_poll_listener(&mut self) -> ApplicationResult<Option<Event<UserEvent>>> {
+        self.listener.try_poll().map_err(ApplicationError::from)
     }
 
     /// Forward event to current active component, if any.
@@ -389,6 +421,14 @@ pub enum PollStrategy {
     ///
     /// This will block if there are not `n` events for up to the timeout specified in `EventListenerCfg`.
     UpTo(usize),
+
+    /// Practially the same as [`UpTo`](PollStrategy::UpTo), only that it does not wait the timeout specified in `EventListenerCfg` if there is at least one event.
+    ///
+    /// This will because of that wait for the first event to be come available for `timeout`, but collect all remaining event without waiting.
+    /// If there are enough event already, never blocks.
+    ///
+    /// This *might* become the default [`UpTo`](PollStrategy::UpTo) behavior in the future.
+    UpToNoWait(usize),
 }
 
 // -- error
@@ -758,6 +798,62 @@ mod test {
             .ok()
             .unwrap();
         assert!(events.len() >= 2);
+    }
+
+    #[test]
+    fn strategy_upto_nowait_should_work() {
+        let mut application: Application<MockComponentId, MockMsg, MockEvent> = Application::init(
+            listener_config_with_tick(Duration::from_secs(60)).poll_timeout(Duration::from_secs(5)),
+        );
+
+        // Mount foo and bar
+        assert!(
+            application
+                .mount(
+                    MockComponentId::InputFoo,
+                    Box::new(MockFooInput::default()),
+                    vec![]
+                )
+                .is_ok()
+        );
+        assert!(
+            application
+                .mount(
+                    MockComponentId::InputBar,
+                    Box::new(MockBarInput::default()),
+                    vec![
+                        Sub::new(SubEventClause::Tick, SubClause::Always),
+                        Sub::new(
+                            // NOTE: won't be thrown, since requires focus
+                            SubEventClause::Keyboard(KeyEvent::from(Key::Enter)),
+                            SubClause::HasAttrValue(
+                                MockComponentId::InputBar,
+                                Attribute::Focus,
+                                AttrValue::Flag(true)
+                            )
+                        )
+                    ]
+                )
+                .is_ok()
+        );
+        // Active FOO
+        assert!(application.active(&MockComponentId::InputFoo).is_ok());
+
+        let before = Instant::now();
+
+        assert_eq!(
+            application
+                .tick(PollStrategy::UpToNoWait(5))
+                .ok()
+                .unwrap()
+                .as_slice(),
+            &[MockMsg::FooSubmit(String::new()), MockMsg::BarTick]
+        );
+
+        let time_taken = Instant::now().duration_since(before);
+
+        // messages should be available, so "UpToNoWait" should not block again after the first event
+        assert!(time_taken < Duration::from_secs(5));
     }
 
     #[test]